refactor: standardize `projects` routes, add read scope handling, and implement repository method for project listing
